dobot/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydobot==1.3.2
pyserial==3.4
Werkzeug==2.2.3
